Move calculation logic to simulation
