+++ tests/core/test_cycles.py
@@ -5,7 +5,7 @@ from nose.tools import raises
 
 from openfisca_core import periods
 from openfisca_core.periods import MONTH
-from openfisca_core.formulas import CycleError
+from openfisca_core.simulations import CycleError
 from openfisca_core.variables import Variable
 
 from openfisca_country_template import CountryTaxBenefitSystem
